+++ core/src/main/kotlin/dev/kensa/parse/ParseTreeEvent.kt
@@ -55,6 +55,8 @@ sealed class LocatedEvent(val location: Location) : Event() {
 
     class Identifier(location: Location, val name: String) : LocatedEvent(location)
 
+    class ConstantReference(location: Location, val name: String, val hint: String) : LocatedEvent(location)
+
     class Operator(location: Location, val text: String) : LocatedEvent(location)
     class Parameter(location: Location, val name: String) : LocatedEvent(location)
     class Method(location: Location, val name: String) : LocatedEvent(location)